+++ ovp_users/models/user.py
@@ -49,6 +49,7 @@ class User(AbstractBaseUser, PermissionsMixin):
   slug = models.SlugField(_('Slug'), max_length=100, null=True, blank=True, unique=True)
   phone = models.CharField(_('Phone'), max_length=30, null=True, blank=True)
   avatar = models.ForeignKey('ovp_uploads.UploadedImage', blank=False, null=True, related_name='avatar_user', verbose_name=_('avatar'))
+  public = models.BooleanField(_('Public'), default=True)
 
   locale = models.CharField(_('Locale'), max_length=8, null=False, blank=True, default='en')