+++ ovp_users/views/__init__.py
@@ -1,4 +1,5 @@
 from ovp_users.views.user import UserResourceViewSet
+from ovp_users.views.user import PublicUserResourceViewSet
 from ovp_users.views.password_recovery import RecoveryTokenViewSet
 from ovp_users.views.password_recovery import RecoverPasswordViewSet
 from ovp_users.views.messages import UserMessageViewSet