+++ CHANGES.rst
@@ -166,3 +166,4 @@ v1.1.3
 v1.1.4[unreleased]
 -----------
 * Add dynamic profile models
+* Add slug field. Default is shortuuid