+++ CHANGES.rst
@@ -167,3 +167,4 @@ v1.1.4[unreleased]
 -----------
 * Add dynamic profile models
 * Add slug field. Default is shortuuid
+* Add User.public field.